Update personal keymap
